Use a mapped/offset array to store Chebyshev coefficients
The Chebyshev coefficients associated with the induced velocity on
the plate are complex valued.  Most of the time, we only need the
imaginary component of the coefficients.  Instead of allocating a
new array to store just the imaginary components, we create a wrapper
that calls `imag` on demand.  The wrapper can also apply an indexing
offset so that the starting index of the array begins at 0, making
it more consistent with the derivations.