feature: functions to compute plate circulation
The `bound_circulation` function used to compute the bound vortex sheet
strength, which is actually the derivative of circulation with respect
to the sheet arc length.  Those functions were renamed to `strength`,
and `bound_circulation[!]` now computes the amount of circulation
between the trailing edge of the sheet (s = -1) and any point along the
sheet.