refactor: use traits to dispatch induced property
Previously, the `induce_velocity(!)` methods were dispatched directly
based on just the types of the target and source.  This made it
difficult to write generic code without encountering method ambiguities.
Now, the specific method that is called depends on which traits a vortex
element chooses to adopt (by implementing the required functions).
This should make it a lot more flexible.