refactor: reimplement the property macro
So not really refactoring... The complexity definitely increased,
but the propery macro is now a lot more flexible.  Whereas the
functions previously generated had a fixed argument order, i.e.

    induce_property(target, target_attrs, source, source_attrs)

users can now specify the function signature explicitly:

    induce_property(targ::Target, t, src::Source)

where the type annotations are used to specify the roles of the
different arguments.  If no targets are specified, then we assume that
this is an intrinsic property (such as position).  For these types of
properties, the user can also specify whether or not it is a
reducible/foldable property (for aggregating the results).  For
example:

    @property begin
        signature = circulation(s::Source)
        reduce = (+)
        stype = Complex128
    end

will create the `circulation(group)` function that recursively calls
itself on the elements of `group` until it reaches singleton vortex
elements, which require their own `circulation` function to be defined.
Without `reduce = (+)`, calling `circulation` on an array of vortex
elements should return an array of circulations.  Need better docs...