fix: clean up `property` macro and define `force` using it
Force is computed using the time derivative of linear impulse.  We can
sum up the rate of change of impulse of all individual vortex elements
to find the total change in impulse.  The `property` macro
automatically takes care of recursively dispatching and distributing
the function calls among groups of vortex elements.