undecided: compute pressure on the plate without finite differences
One of the terms we need to compute pressure on the plate is the rate
of change of the bound circulation.  Previously, we have been
computing it using finite-differences, i.e. computing the bound
circulation distribution at every time-step, then applying a backward
difference to determine its rate of change.  Instead, this change
directly computes the derivative of the bound circulation (caveat:
the vorticity flux due to newly created vortex elements are determined
with finite differences)